Support title format modifications (#18)
* chore: add "to"

* docs: small change to example

* docs: provide longest label in description

* feat: reinstitute support for titleformat settings

Fixes #17. Setting this up on the user side via beamer templates turned
out to be too much of a hassle for users. It seems that the
implementation here works better out of the box compared to what was
originally the case in the metropolis theme, and we don't have to rely
on patching commands as the previous implementation did, so this seems
like a reasonable feature to add back support for.