feat: redesign section and subsection pages  (#30)
Redesign section and subsection pages so that section titles are aligned, which 
helps avoid having titles jump if section and subsection pages are viewed
in sequence. The alignment of subtitles on section pages with
the progressbar layout have also been modified to look better. Finally,
section pages have their content raised slightly in both the progressbar
and simple templates.

---------

Co-authored-by: Johan Larsson <jolars@posteo.com>